put fused_moe back in fp8.py
